refactor: fsd 스타일로 폴더 구조 변경
