+++ client-toss/src/app/config/router.tsx
@@ -1,9 +1,9 @@
-import { Dashboard } from "@/views/dashboard/Dashboard";
+import { Dashboard } from "@/views/dashboard";
 import { Drawing } from "@/views/drawing";
 import { HomeView } from "@/views/home";
 import { Memorize } from "@/views/memorize";
 import { SubmittedView } from "@/views/submitted";
-import { Ranking } from "@/views/ranking/Ranking";
+import { Ranking } from "@/views/ranking";
 import { createBrowserRouter } from "react-router-dom";
 
 export const router = createBrowserRouter([
