+++ client-toss/src/entities/podium/index.ts
@@ -1,0 +1,1 @@
+export { default as DailyPodium } from "./ui/DailyPodium";
