+++ client-toss/src/shared/ui/arcScoreBar/index.ts
@@ -1,0 +1,1 @@
+export { default as ArcScoreBar } from "./ArcScoreBar";
