+++ client-toss/src/views/ranking/index.ts
@@ -1,0 +1,1 @@
+export { default as Ranking } from "./ui/Ranking";
